# restaurantapi-main/openapi/swagger-latest-team12.json
{
  "swagger": "2.0",
  "info": {
    "version": "2025-04-15T08:32:20Z",
    "title": "tm3-restaurant_api-sam5"
  },
  "host": "3rh91vhdtj.execute-api.eu-west-3.amazonaws.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT Authorization header using the Bearer scheme. Example: 'Authorization: Bearer {token}'"
    }
  },
  "paths": {
...
    "/feedbacks": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "returns list of feedbacks",
        "description": "This endpoint returns a list of all feedbacks with pagination and filtering options",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter feedbacks by type",
            "required": false,
            "type": "string",
            "enum": [
              "Cuisine",
              "Service",
              "Ambiance",
              "Value",
              "Overall"
            ]
          },
          {
            "name": "locationId",
            "in": "query",
            "description": "Filter feedbacks by location ID",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "required": false,
            "type": "integer",
            "default": 0
          },
          {
            "name": "size",
            "in": "query",
            "description": "Number of items per page",
            "required": false,
            "type": "integer",
            "default": 20
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "content": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "rate": {
                        "type": "string"
                      },
                      "comment": {
                        "type": "string"
                      },
                      "userName": {
                        "type": "string"
                      },
                      "userAvatarUrl": {
                        "type": "string"
                      },
                      "date": {
                        "type": "string"
                      },
                      "type": {
                        "type": "string"
                      },
                      "locationId": {
                        "type": "string"
                      }
                    }
                  }
                },
                "totalPages": {
                  "type": "integer"
                },
                "totalElements": {
                  "type": "integer"
                },
                "size": {
                  "type": "integer"
                },
                "number": {
                  "type": "integer"
                },
                "sort": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "direction": {
                        "type": "string"
                      },
                      "property": {
                        "type": "string"
                      },
                      "ignoreCase": {
                        "type": "boolean"
                      },
                      "nullHandling": {
                        "type": "string"
                      },
                      "ascending": {
                        "type": "boolean"
                      }
                    }
                  }
                },
                "first": {
                  "type": "boolean"
                },
                "last": {
                  "type": "boolean"
                },
                "numberOfElements": {
                  "type": "integer"
                },
                "pageable": {
                  "type": "object",
                  "properties": {
                    "offset": {
                      "type": "integer"
                    },
                    "sort": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "direction": {
                            "type": "string"
                          },
                          "property": {
                            "type": "string"
                          },
                          "ignoreCase": {
                            "type": "boolean"
                          },
                          "nullHandling": {
                            "type": "string"
                          },
                          "ascending": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "pageNumber": {
                      "type": "integer"
                    },
                    "pageSize": {
                      "type": "integer"
                    },
                    "paged": {
                      "type": "boolean"
                    },
                    "unpaged": {
                      "type": "boolean"
                    }
                  }
                },
                "empty": {
                  "type": "boolean"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      },
      "post": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "creates a new feedback",
        "description": "This endpoint allows users to submit a new feedback",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Feedback object that needs to be submitted",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "rate",
                "type",
                "locationId"
              ],
              "properties": {
                "rate": {
                  "type": "string",
                  "description": "Rating value (1-5)"
                },
                "comment": {
                  "type": "string",
                  "description": "User's comment"
                },
                "type": {
                  "type": "string",
                  "description": "Type of feedback",
                  "enum": [
                    "Cuisine",
                    "Service",
                    "Ambiance",
                    "Value",
                    "Overall"
                  ]
                },
                "locationId": {
                  "type": "string",
                  "description": "ID of the restaurant location"
                },
                "reservationId": {
                  "type": "string",
                  "description": "ID of the reservation (optional)"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Feedback created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Location not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
    "/feedbacks/{id}": {
      "get": {
        "tags": [
          "Feedbacks"
        ],
        "summary": "returns a specific feedback by ID",
        "description": "This endpoint returns detailed information about a specific feedback identified by its ID",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Unique identifier of the feedback",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Feedback details successfully retrieved",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string"
                },
                "reservationId": {
                  "type": "string"
                },
                "customerId": {
                  "type": "string"
                },
                "customerName": {
                  "type": "string"
                },
                "rating": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "comment": {
                  "type": "string"
                },
                "serviceRating": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "foodRating": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "ambienceRating": {
                  "type": "number",
                  "minimum": 1,
                  "maximum": 5
                },
                "visitDate": {
                  "type": "string",
                  "format": "date"
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "locationId": {
                  "type": "string"
                },
                "locationName": {
                  "type": "string"
                },
                "photos": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "response": {
                  "type": "object",
                  "properties": {
                    "text": {
                      "type": "string"
                    },
                    "respondentName": {
                      "type": "string"
                    },
                    "respondentRole": {
                      "type": "string"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            },
            "examples": {
              "application/json": {
                "id": "feedback-123",
                "reservationId": "res-456",
                "customerId": "user-789",
                "customerName": "John Doe",
                "rating": 4.5,
                "comment": "We had a wonderful dining experience. The food was excellent and the service was attentive. The only small issue was that it was a bit noisy, but overall we had a great time and will definitely return.",
                "serviceRating": 5,
                "foodRating": 5,
                "ambienceRating": 3.5,
                "visitDate": "2024-04-15",
                "createdAt": "2024-04-16T10:30:00Z",
                "locationId": "loc-101",
                "locationName": "Green & Tasty Downtown",
                "photos": [
                  "https://green-and-tasty.s3.eu-central-1.amazonaws.com/feedback/img123.jpg",
                  "https://green-and-tasty.s3.eu-central-1.amazonaws.com/feedback/img456.jpg"
                ],
                "response": {
                  "text": "Thank you for your wonderful feedback, John! We're delighted you enjoyed the food and service. We're working on improving the acoustics to address the noise level. We look forward to welcoming you back soon!",
                  "respondentName": "Sarah Johnson",
                  "respondentRole": "Restaurant Manager",
                  "createdAt": "2024-04-17T09:15:00Z"
                }
              }
            },
            "headers": {
              "Access-Control-Allow-Origin": {
                "type": "string"
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication token is missing or invalid",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Unauthorized: Valid authentication token is required"
              }
            }
          },
          "404": {
            "description": "Feedback not found",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            },
            "examples": {
              "application/json": {
                "message": "Feedback with ID 'feedback-123' not found"
              }
            }
          }
        },
        "x-amazon-apigateway-integration": {
          "httpMethod": "POST",
          "uri": "arn:aws:apigateway:eu-west-3:lambda:path/2015-03-31/functions/arn:aws:lambda:eu-west-3:089718700404:function:tm3-api-handler-sam5/invocations",
          "responses": {
            "default": {
              "statusCode": "200",
              "responseParameters": {
                "method.response.header.Access-Control-Allow-Origin": "'*'"
              }
            }
          },
          "passthroughBehavior": "when_no_match",
          "type": "aws_proxy"
        }
      }
    },
...
  },
...
}